health and manabar

+++ src/data.c
@@ -40,7 +40,7 @@ static void entity_push_data(Entity* entity, u32 offset)
     data.buffer[offset++] = entity->position.y;
     data.buffer[offset++] = entity->position.z;
     data.buffer[offset++] = entity->scale;
-    data.buffer[offset++] = (f32)entity->health / entity->max_health;
+    data.buffer[offset++] = entity->health / entity->max_health;
     f32 theta = atan(entity->facing.y / entity->facing.x) + (entity->facing.x >= 0 ? 0 : PI);
 
     f32 dif = theta + PI - camera.yaw;
